getting path to work

+++ external/atari_wrappers.py
@@ -123,7 +123,7 @@ class WarpFrame(gym.ObservationWrapper):
         gym.ObservationWrapper.__init__(self, env)
         self.width = 84
         self.height = 84
-        self.observation_space = spaces.Box(low=0, high=255, shape=(self.height, self.width, 1))
+        self.observation_space = spaces.Box(low=0, high=255, shape=(self.height, self.width, 1), dtype=np.uint8)
 
     def _observation(self, frame):
         frame = cv2.cvtColor(frame, cv2.COLOR_RGB2GRAY)
@@ -144,7 +144,7 @@ class FrameStack(gym.Wrapper):
         self.k = k
         self.frames = deque([], maxlen=k)
         shp = env.observation_space.shape
-        self.observation_space = spaces.Box(low=0, high=255, shape=(shp[0], shp[1], shp[2] * k))
+        self.observation_space = spaces.Box(low=0, high=255, shape=(shp[0], shp[1], shp[2] * k), dtype=np.uint8)
 
     def _reset(self):
         ob = self.env.reset()
@@ -196,14 +196,20 @@ def wrap_deepmind(env, episode_life=True, clip_rewards=True, frame_stack=False, 
     """
     if episode_life:
         env = EpisodicLifeEnv(env)
+    
     if 'FIRE' in env.unwrapped.get_action_meanings():
         env = FireResetEnv(env)
+    
     env = WarpFrame(env)
+    
     if scale:
         env = ScaledFloatFrame(env)
+    
     if clip_rewards:
         env = ClipRewardEnv(env)
+    
     if frame_stack:
         env = FrameStack(env, 4)
+    
     return env
